finish move & change callback signature
