+++ src/renderer/components/content-types/board/Board.css
@@ -1,5 +1,5 @@
 .Board {
   position: relative;
-  background-image: url('../../images/grid.svg');
+  background-image: url('../../../images/grid.svg');
   background-blend-mode: exclusion;
 }
