+++ src/renderer/components/content-types/board/Board.tsx
@@ -5,9 +5,9 @@ import uuid from 'uuid/v4'
 
 import { Handle } from 'hypermerge'
 import { ContextMenuTrigger } from 'react-contextmenu'
-import { ContentProps } from '../Content'
-import ContentTypes from '../../ContentTypes'
-import { parseDocumentLink, PushpinUrl, isPushpinUrl } from '../../ShareLink'
+import { ContentProps } from '../../Content'
+import ContentTypes from '../../../ContentTypes'
+import { parseDocumentLink, PushpinUrl, isPushpinUrl } from '../../../ShareLink'
 import { BoardDoc } from '.'
 import BoardCard from './BoardCard'
 import BoardContextMenu from './BoardContextMenu'
