+++ src/renderer/components/content-types/defaults/DefaultInList.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import ContentTypes from '../../ContentTypes'
-import { ContentProps } from '../Content'
-import { useDocument } from '../../Hooks'
-import Badge from '../Badge'
+import ContentTypes from '../../../ContentTypes'
+import { ContentProps } from '../../Content'
+import { useDocument } from '../../../Hooks'
+import Badge from '../../Badge'
 
 interface Doc {
   title?: string
