+++ src/renderer/components/content-types/workspace/Share.tsx
@@ -2,21 +2,21 @@ import React, { useState } from 'react'
 import Debug from 'debug'
 import classnames from 'classnames'
 
-import Content from '../Content'
-import Popover from '../Popover'
+import Content from '../../Content'
+import Popover from '../../Popover'
 
-import { createDocumentLink, HypermergeUrl } from '../../ShareLink'
+import { createDocumentLink, HypermergeUrl } from '../../../ShareLink'
 import { Doc as WorkspaceDoc } from './Workspace'
 import { ContactDoc } from '../contact'
-import Label from '../Label'
-import Badge from '../Badge'
-import { useDocument, useRepo } from '../../Hooks'
-import Text from '../Text'
-import SecondaryText from '../SecondaryText'
+import Label from '../../Label'
+import Badge from '../../Badge'
+import { useDocument, useRepo } from '../../../Hooks'
+import Text from '../../Text'
+import SecondaryText from '../../SecondaryText'
 import ActionListItem from './ActionListItem'
-import ListMenuItem, { Stretch } from '../ListMenuItem'
-import ListMenuSection from '../ListMenuSection'
-import ListMenu from '../ListMenu'
+import ListMenuItem, { Stretch } from '../../ListMenuItem'
+import ListMenuSection from '../../ListMenuSection'
+import ListMenu from '../../ListMenu'
 import './Share.css'
 
 const log = Debug('pushpin:share')
