+++ src/renderer/components/content-types/workspace/TitleBar.tsx
@@ -3,13 +3,13 @@ import { clipboard } from 'electron'
 
 import Dropdown, { DropdownContent, DropdownTrigger } from 'react-simple-dropdown/dropdown'
 import Omnibox from './Omnibox'
-import Content from '../Content'
+import Content from '../../Content'
 import Authors from './Authors'
 import Share from './Share'
-import { HypermergeUrl, PushpinUrl } from '../../ShareLink'
+import { HypermergeUrl, PushpinUrl } from '../../../ShareLink'
 
 import './TitleBar.css'
-import { useDocument, useEvent } from '../../Hooks'
+import { useDocument, useEvent } from '../../../Hooks'
 
 export interface Props {
   hypermergeUrl: HypermergeUrl
